requirements.txt
# ── Core API framework ────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2

# ── OpenCV and related image processing (legacy / fallback provider) ──────
opencv-python-headless==4.8.1.78
imutils==0.5.4
numpy==1.26.4
Pillow==10.4.0

# ── Flask and its extensions (prediction / model API) ────────────────────
Flask==3.0.3
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0

# ── Machine learning libraries ────────────────────────────────────────────
scikit-learn==1.5.2
pandas==2.2.3
matplotlib==3.9.2

# ── HuggingFace / local model provider ───────────────────────────────────
transformers==4.44.2
# torch is optional — comment out if running CPU-only in CI to reduce install time
# torch==2.4.0

# ── MongoDB ───────────────────────────────────────────────────────────────
pymongo==4.8.0

# ── Authentication and security ───────────────────────────────────────────
bcrypt==4.2.0
python-dotenv==1.0.1
PyJWT==2.9.0

# ── HTTP client ───────────────────────────────────────────────────────────
requests==2.32.3
httpx==0.27.2

# ── Data handling ─────────────────────────────────────────────────────────
jsonschema==4.23.0

# ── Scheduling ────────────────────────────────────────────────────────────
APScheduler==3.10.4

# ── Testing ───────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2  
